+++ src/main/java/com/nisovin/shopkeepers/shopkeeper/admin/regular/SKRegularAdminShopkeeper.java
@@ -12,21 +12,21 @@ import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperCreateException;
 import com.nisovin.shopkeepers.api.shopkeeper.TradingRecipe;
 import com.nisovin.shopkeepers.api.shopkeeper.admin.regular.RegularAdminShopkeeper;
-import com.nisovin.shopkeepers.api.shopkeeper.offers.TradingOffer;
+import com.nisovin.shopkeepers.api.shopkeeper.offers.TradeOffer;
 import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.debug.DebugOptions;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.shopkeeper.SKDefaultShopTypes;
 import com.nisovin.shopkeepers.shopkeeper.admin.AbstractAdminShopkeeper;
-import com.nisovin.shopkeepers.shopkeeper.offers.SKTradingOffer;
+import com.nisovin.shopkeepers.shopkeeper.offers.SKTradeOffer;
 import com.nisovin.shopkeepers.util.Log;
 import com.nisovin.shopkeepers.util.Validate;
 
 public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements RegularAdminShopkeeper {
 
-	// Can contain multiple offers for a specific type of item:
-	private final List<TradingOffer> offers = new ArrayList<>();
-	private final List<TradingOffer> offersView = Collections.unmodifiableList(offers);
+	// There can be multiple different offers for the same kind of item:
+	private final List<TradeOffer> offers = new ArrayList<>();
+	private final List<TradeOffer> offersView = Collections.unmodifiableList(offers);
 
 	// Kept in sync with offers:
 	private final List<TradingRecipe> recipes = new ArrayList<>();
@@ -66,12 +66,10 @@ public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements
 	protected void loadFromSaveData(ConfigurationSection configSection) throws ShopkeeperCreateException {
 		super.loadFromSaveData(configSection);
 		// Load offers:
-		List<SKTradingOffer> offers = SKTradingOffer.loadFromConfig(configSection, "recipes", "Shopkeeper " + this.getId());
-		List<SKTradingOffer> migratedOffers = SKTradingOffer.migrateItems(offers, "Shopkeeper " + this.getId());
+		List<SKTradeOffer> offers = SKTradeOffer.loadFromConfig(configSection, "recipes", "Shopkeeper " + this.getId());
+		List<SKTradeOffer> migratedOffers = SKTradeOffer.migrateItems(offers, "Shopkeeper " + this.getId());
 		if (offers != migratedOffers) {
-			Log.debug(DebugOptions.itemMigrations,
-					() -> "Shopkeeper " + this.getId() + ": Migrated trading offer items."
-			);
+			Log.debug(DebugOptions.itemMigrations, () -> "Shopkeeper " + this.getId() + ": Migrated trade offer items.");
 			this.markDirty();
 		}
 		this._setOffers(migratedOffers);
@@ -81,7 +79,7 @@ public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements
 	public void save(ConfigurationSection configSection) {
 		super.save(configSection);
 		// Save offers:
-		SKTradingOffer.saveToConfig(configSection, "recipes", this.getOffers());
+		SKTradeOffer.saveToConfig(configSection, "recipes", this.getOffers());
 	}
 
 	@Override
@@ -97,7 +95,7 @@ public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements
 	// OFFERS:
 
 	@Override
-	public List<TradingOffer> getOffers() {
+	public List<TradeOffer> getOffers() {
 		return offersView;
 	}
 
@@ -113,31 +111,31 @@ public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements
 	}
 
 	@Override
-	public void setOffers(List<TradingOffer> offers) {
+	public void setOffers(List<TradeOffer> offers) {
 		Validate.notNull(offers, "Offers is null!");
 		Validate.noNullElements(offers, "Offers contains null elements!");
 		this._setOffers(offers);
 		this.markDirty();
 	}
 
-	private void _setOffers(List<? extends TradingOffer> offers) {
+	private void _setOffers(List<? extends TradeOffer> offers) {
 		assert offers != null && !offers.contains(null);
 		this._clearOffers();
 		this._addOffers(offers);
 	}
 
 	@Override
-	public void addOffer(TradingOffer offer) {
+	public void addOffer(TradeOffer offer) {
 		Validate.notNull(offer, "Offer is null!");
 		this._addOffer(offer);
 		this.markDirty();
 	}
 
-	private void _addOffer(TradingOffer offer) {
+	private void _addOffer(TradeOffer offer) {
 		assert offer != null;
 		offers.add(offer);
 		if (offer instanceof TradingRecipe) {
-			// SKTradingOffer extends SKTradingRecipe and reports to not be out-of-stock.
+			// SKTradeOffer extends SKTradingRecipe and reports to not be out-of-stock.
 			recipes.add((TradingRecipe) offer);
 		} else {
 			recipes.add(ShopkeepersAPI.createTradingRecipe(offer.getResultItem(), offer.getItem1(), offer.getItem2(), false));
@@ -145,14 +143,14 @@ public class SKRegularAdminShopkeeper extends AbstractAdminShopkeeper implements
 	}
 
 	@Override
-	public void addOffers(List<TradingOffer> offers) {
+	public void addOffers(List<TradeOffer> offers) {
 		Validate.notNull(offers, "Offers is null!");
 		Validate.noNullElements(offers, "Offers contains null elements!");
 		this._addOffers(offers);
 		this.markDirty();
 	}
 
-	private void _addOffers(List<? extends TradingOffer> offers) {
+	private void _addOffers(List<? extends TradeOffer> offers) {
 		assert offers != null && !offers.contains(null);
 		offers.forEach(this::_addOffer);
 	}