+++ src/main/java/com/nisovin/shopkeepers/shopkeeper/player/trade/SKTradingPlayerShopkeeper.java
@@ -11,14 +11,14 @@ import org.bukkit.inventory.ItemStack;
 import com.nisovin.shopkeepers.api.ShopkeepersAPI;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperCreateException;
 import com.nisovin.shopkeepers.api.shopkeeper.TradingRecipe;
-import com.nisovin.shopkeepers.api.shopkeeper.offers.TradingOffer;
+import com.nisovin.shopkeepers.api.shopkeeper.offers.TradeOffer;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopCreationData;
 import com.nisovin.shopkeepers.api.shopkeeper.player.trade.TradingPlayerShopkeeper;
 import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.debug.DebugOptions;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.shopkeeper.SKDefaultShopTypes;
-import com.nisovin.shopkeepers.shopkeeper.offers.SKTradingOffer;
+import com.nisovin.shopkeepers.shopkeeper.offers.SKTradeOffer;
 import com.nisovin.shopkeepers.shopkeeper.player.AbstractPlayerShopkeeper;
 import com.nisovin.shopkeepers.util.ItemCount;
 import com.nisovin.shopkeepers.util.ItemUtils;
@@ -27,8 +27,9 @@ import com.nisovin.shopkeepers.util.Validate;
 
 public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implements TradingPlayerShopkeeper {
 
-	private final List<TradingOffer> offers = new ArrayList<>();
-	private final List<TradingOffer> offersView = Collections.unmodifiableList(offers);
+	// There can be multiple different offers for the same kind of item:
+	private final List<TradeOffer> offers = new ArrayList<>();
+	private final List<TradeOffer> offersView = Collections.unmodifiableList(offers);
 
 	/**
 	 * Creates a not yet initialized {@link SKTradingPlayerShopkeeper} (for use in sub-classes).
@@ -67,8 +68,8 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	protected void loadFromSaveData(ConfigurationSection configSection) throws ShopkeeperCreateException {
 		super.loadFromSaveData(configSection);
 		// Load offers:
-		List<SKTradingOffer> offers = SKTradingOffer.loadFromConfig(configSection, "offers", "Shopkeeper " + this.getId());
-		List<SKTradingOffer> migratedOffers = SKTradingOffer.migrateItems(offers, "Shopkeeper " + this.getId());
+		List<SKTradeOffer> offers = SKTradeOffer.loadFromConfig(configSection, "offers", "Shopkeeper " + this.getId());
+		List<SKTradeOffer> migratedOffers = SKTradeOffer.migrateItems(offers, "Shopkeeper " + this.getId());
 		if (offers != migratedOffers) {
 			Log.debug(DebugOptions.itemMigrations,
 					() -> "Shopkeeper " + this.getId() + ": Migrated trading offer items."
@@ -82,7 +83,7 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	public void save(ConfigurationSection configSection) {
 		super.save(configSection);
 		// Save offers:
-		SKTradingOffer.saveToConfig(configSection, "offers", this.getOffers());
+		SKTradeOffer.saveToConfig(configSection, "offers", this.getOffers());
 	}
 
 	@Override
@@ -94,7 +95,7 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	public List<TradingRecipe> getTradingRecipes(Player player) {
 		List<TradingRecipe> recipes = new ArrayList<>();
 		List<ItemCount> containerItems = this.getItemsFromContainer();
-		for (TradingOffer offer : this.getOffers()) {
+		for (TradeOffer offer : this.getOffers()) {
 			ItemStack resultItem = offer.getResultItem();
 			assert !ItemUtils.isEmpty(resultItem);
 			int itemAmountInContainer = 0;
@@ -118,13 +119,13 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	// OFFERS:
 
 	@Override
-	public List<TradingOffer> getOffers() {
+	public List<TradeOffer> getOffers() {
 		return offersView;
 	}
 
 	// Note: There might be multiple offers involving this item, this only returns the first one it finds.
-	public TradingOffer getOffer(ItemStack tradedItem) {
-		for (TradingOffer offer : this.getOffers()) {
+	public TradeOffer getOffer(ItemStack tradedItem) {
+		for (TradeOffer offer : this.getOffers()) {
 			if (ItemUtils.isSimilar(offer.getResultItem(), tradedItem)) {
 				return offer;
 			}
@@ -132,8 +133,8 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 		return null;
 	}
 
-	public TradingOffer getOffer(TradingRecipe tradingRecipe) {
-		for (TradingOffer offer : this.getOffers()) {
+	public TradeOffer getOffer(TradingRecipe tradingRecipe) {
+		for (TradeOffer offer : this.getOffers()) {
 			if (offer.areItemsEqual(tradingRecipe)) {
 				return offer;
 			}
@@ -152,40 +153,40 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	}
 
 	@Override
-	public void setOffers(List<TradingOffer> offers) {
+	public void setOffers(List<TradeOffer> offers) {
 		Validate.notNull(offers, "Offers is null!");
 		Validate.noNullElements(offers, "Offers contains null elements!");
 		this._setOffers(offers);
 		this.markDirty();
 	}
 
-	private void _setOffers(List<? extends TradingOffer> offers) {
+	private void _setOffers(List<? extends TradeOffer> offers) {
 		assert offers != null && !offers.contains(null);
 		this._clearOffers();
 		this._addOffers(offers);
 	}
 
 	@Override
-	public void addOffer(TradingOffer offer) {
+	public void addOffer(TradeOffer offer) {
 		Validate.notNull(offer, "Offer is null!");
 		this._addOffer(offer);
 		this.markDirty();
 	}
 
-	private void _addOffer(TradingOffer offer) {
+	private void _addOffer(TradeOffer offer) {
 		assert offer != null;
 		offers.add(offer);
 	}
 
 	@Override
-	public void addOffers(List<TradingOffer> offers) {
+	public void addOffers(List<TradeOffer> offers) {
 		Validate.notNull(offers, "Offers is null!");
 		Validate.noNullElements(offers, "Offers contains null elements!");
 		this._addOffers(offers);
 		this.markDirty();
 	}
 
-	private void _addOffers(List<? extends TradingOffer> offers) {
+	private void _addOffers(List<? extends TradeOffer> offers) {
 		assert offers != null && !offers.contains(null);
 		offers.forEach(this::_addOffer);
 	}